Added stats fields and started on update
Currently the update page is piggybacked off the Create page when you specify an id in the route. Still need to post the data to the Update route if it's an update and load the current images. Probably add an option to remove the current ones which deletes the files