Design changes
Updated the design of the kiosk page. The sidepanel is now based on the page width opposed to it's height, the return button has been relocated to the bottom of the page and the title of the page you are on remains on top at all times.
TODO- add button to toggle sidebar